modulized outdoor_main
